chore: move electronics-library-dev to electronics-library; minor tweaks; remove package-lock (since we're using yarn)
